refactor: extract ui tabs
